:truck: Move Navbar, Sidebar inside Navigation
