+++ src/components/Navigation/Navigation.js
@@ -1,8 +1,9 @@
 import React from "react";
-import NavBar from "../NavBar/NavBar";
-import GuestNavBar from "../NavBar/GuestNavBar";
+import NavBar from "./components/NavBar/NavBar";
+import GuestNavBar from "./components/NavBar/GuestNavBar";
+import SideBar from "./components/SideBar/SideBar";
 import _ from "lodash";
-import { Redirect } from "react-router-dom";
+
 const Navigation = (props) => {
   const { data, signOutUser } = props;
   const logOutUser = () => {
@@ -10,14 +11,16 @@ const Navigation = (props) => {
     setTimeout(() => {
       window.location.reload();
     }, 20);
-    //
   };
   return (
     <div>
       {data.userData === null ? (
         <GuestNavBar />
       ) : (
-        <NavBar username={data.userData.username} logOutUser={logOutUser} />
+        <>
+          <SideBar />
+          <NavBar username={data.userData.username} logOutUser={logOutUser} />
+        </>
       )}
     </div>
   );
