+++ src/components/Navigation/components/NavBar/default_image.jpg
